NerdleApp/back/requirements.txt
Flask~=2.3.2
numpy~=1.24.4
Flask-Cors~=3.0.10
python-dotenv~=1.0.0
cryptography~=41.0.3
PyJWT~=2.8.0
progress~=1.6
